refactor(db): Split collection_ops.rs into directory module
Split collection_ops.rs (1,342 lines) into 5 focused files:
- mod.rs: imports, load_collections(), reload_cache(), shared helpers
- create.rs: next_collection_short_id(), create_collection*()
- delete.rs: delete_collection*(), truncate_collection*()
- lookup.rs: list/get/has/find/snapshot collection methods
- version.rs: set_active_version, get/delete collection versions
